Refactor/cleanup code according to code review
