+++ src/main/java/com/assemblyrobot/simulator/core/metrics/MetricsCollectorType.java
@@ -17,7 +17,6 @@ public enum MetricsCollectorType {
    * @throws MetricsCollectorTypeNotRegisteredError If no such organisational identifier exists.
    */
 
-  //TODO: Check cases to match classnames
   public static MetricsCollectorType getByClass(@NonNull String className) throws MetricsCollectorTypeNotRegisteredError{
     return switch (className) {
       case "com.assemblyrobot.simulator.core.Engine" -> ENGINE;
