+++ src/main/java/com/assemblyrobot/simulator/system/components/Tracker.java
@@ -1,6 +1,5 @@
 package com.assemblyrobot.simulator.system.components;
 
-import com.assemblyrobot.simulator.core.metrics.MaterialStationData;
 import java.util.ArrayList;
 import lombok.Getter;
 import lombok.NonNull;
