+++ src/main/java/com/assemblyrobot/simulator/system/controllers/StageController.java
@@ -4,7 +4,7 @@ import com.assemblyrobot.simulator.core.clock.Clock;
 import com.assemblyrobot.simulator.core.events.Event;
 import com.assemblyrobot.simulator.core.events.EventQueue;
 import com.assemblyrobot.simulator.core.events.EventType;
-import com.assemblyrobot.simulator.core.metrics.MaterialStationData;
+import com.assemblyrobot.simulator.system.components.MaterialStationData;
 import com.assemblyrobot.simulator.core.metrics.MetricsCollector;
 import com.assemblyrobot.simulator.system.components.Material;
 import com.assemblyrobot.simulator.system.components.Tracker;
